Separate hooks examples into different files
